migrate shop screen to owo-ui, migrate config to owo-config, remove old coin stack experiment, use NbtKey for currency items, add basic piggy bank block

+++ src/main/java/com/glisco/numismaticoverhaul/network/ShopScreenHandlerRequestC2SPacket.java
@@ -3,7 +3,7 @@ package com.glisco.numismaticoverhaul.network;
 import com.glisco.numismaticoverhaul.block.ShopScreenHandler;
 import io.wispforest.owo.network.ServerAccess;
 
-public record ShopScreenHandlerRequestC2SPacket(Action action, int value) {
+public record ShopScreenHandlerRequestC2SPacket(Action action, long value) {
 
     public ShopScreenHandlerRequestC2SPacket(Action action) {
         this(action, 0);
@@ -11,7 +11,7 @@ public record ShopScreenHandlerRequestC2SPacket(Action action, int value) {
 
     public static void handle(ShopScreenHandlerRequestC2SPacket message, ServerAccess access) {
         final var player = access.player();
-        final int value = message.value();
+        final long value = message.value();
 
         if (!(player.currentScreenHandler instanceof ShopScreenHandler shopHandler)) return;
